chore: user 폴더 구조 변경 및 import 경로 일괄 수정
