+++ build.gradle
@@ -38,6 +38,9 @@ dependencies {
 	// Spring Security OAuth2 Core
 	implementation 'org.springframework.security:spring-security-oauth2-core'
 
+	// JAXB API
+	implementation 'javax.xml.bind:jaxb-api:2.3.1'
+
 	// JWT (JJWT)
 	implementation 'io.jsonwebtoken:jjwt-api:0.11.5'
 	runtimeOnly    'io.jsonwebtoken:jjwt-impl:0.11.5'
