+++ src/main/java/com/kakaotech/ott/ott/aiImage/infrastructure/repositoryImpl/AiImageRepositoryImpl.java
@@ -4,8 +4,7 @@ import com.kakaotech.ott.ott.aiImage.domain.model.AiImage;
 import com.kakaotech.ott.ott.aiImage.infrastructure.entity.AiImageEntity;
 import com.kakaotech.ott.ott.aiImage.domain.repository.AiImageJpaRepository;
 import com.kakaotech.ott.ott.aiImage.domain.repository.AiImageRepository;
-import com.kakaotech.ott.ott.user.domain.model.User;
-import com.kakaotech.ott.ott.user.infrastructure.repository.UserJpaRepository;
+import com.kakaotech.ott.ott.user.domain.repository.UserJpaRepository;
 import com.kakaotech.ott.ott.user.infrastructure.entity.UserEntity;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
