+++ src/main/java/com/kakaotech/ott/ott/like/infrastructure/entity/LikeEntity.java
@@ -40,7 +40,7 @@ public class LikeEntity {
     private Long targetId;
 
     @Column(name = "is_active", nullable = false)
-    private Boolean isActive = true;
+    private Boolean isActive;
 
     @CreatedDate
     @Column(name = "created_at", nullable = false, updatable = false)
