+++ src/main/java/com/kakaotech/ott/ott/user/domain/repository/OAuthTokenJpaRepository.java
@@ -1,4 +1,4 @@
-package com.kakaotech.ott.ott.user.infrastructure.repository;
+package com.kakaotech.ott.ott.user.domain.repository;
 
 import com.kakaotech.ott.ott.user.infrastructure.entity.OAuthTokenEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
